Silence warning about comparing unsigned with < 0 (from Clang 19)

+++ src/debug/symbols-common.c
@@ -484,7 +484,7 @@ static int read_pc_debug_names(FILE *fp, symbol_list_t *list, uint32_t offset)
 
 	if (fseek(fp, 0, SEEK_END) < 0)
 		return 0;
-	if ((filesize = ftell(fp)) < 0)
+	if ((long)(filesize = ftell(fp)) < 0)
 		return 0;
 	if (fseek(fp, 0, SEEK_SET) < 0)
 		return 0;
